requirements.txt
pyyaml
python-dotenv
aiohttp 